Added Catching some error, filtering-error fixed

+++ SQL_Like_Query/HDF5DB.py
@@ -10,28 +10,32 @@ from HDF5Object import HDF5Object
 
 #ToDo:
 #       Testklasse für HDF5DB sowie HDF5Object
-#       
+#       pathlib
 class HDF5DB:
     def __init__(self):
         self.hdf5ObjectList = []
        
     def addObject(self, object):
-        self.hdf5ObjectList.append(object)
+        if(object != None):
+            self.hdf5ObjectList.append(object)
 
     def addObjectsFromPath(self, path):
-        #print("Paths:")
-        for folder in os.listdir(path):
-            
-            for hdf5File in glob(path + "/" + folder + "/*.hdf5"):
-                for erfh5File in glob(path + "/" + folder + "/*.erfh5"):
-                    if (h5py.is_hdf5(hdf5File) and h5py.is_hdf5(erfh5File)):
-                        #print(path + "/" + folder)
-                        self.hdf5ObjectList.append(HDF5Object(path + "/" + folder))
-                        
-            if not (h5py.is_hdf5(hdf5File)):
-                print(hdf5File + " is corrupt!\nThe file has been skipped.")                                                                                                            #"There is one or more corrupt files in " + path + "/" + folder + "!\n" + "The folder has been skipped.")
-            if not (h5py.is_hdf5(erfh5File)):
-                print(erfh5File + " is corrupt!\nThe file has been skipped.")
+        if(os.path.exists(path)):
+            for folder in os.listdir(path):            
+                for hdf5File in glob(path + "/" + folder + "/*.hdf5"):
+                    for erfh5File in glob(path + "/" + folder + "/*.erfh5"):
+                        # Check that only *.hdf5 and *.erfh5 files will be opened
+                        if (h5py.is_hdf5(hdf5File) and h5py.is_hdf5(erfh5File)):
+                            #print(path + "/" + folder)
+                            self.hdf5ObjectList.append(HDF5Object(path + "/" + folder))
+                            
+                    if not (h5py.is_hdf5(hdf5File)):
+                        print(hdf5File + " is corrupt!\nThe file has been skipped.")                                                                                                            #"There is one or more corrupt files in " + path + "/" + folder + "!\n" + "The folder has been skipped.")
+                    if not (h5py.is_hdf5(erfh5File)):
+                        print(erfh5File + " is corrupt!\nThe file has been skipped.")
+            print(str(len(self.hdf5ObjectList)) + " Objects have been added.")
+        else:
+            print("The path " + path + " does not exist! No objects were added!")
 
     def select(self, variable, value, comparisonOperator="="):
         self.selected = []
@@ -66,7 +70,7 @@ class HDF5DB:
                     self.selected.append(obj)
                 elif (variable == "heightRectangle" and value in obj.heightRectangle):
                     self.selected.append(obj)
-                elif (variable == "widthRectangle" and (value == obj.widthRectangle or value in obj.widthRectangle)):#np.isin(value, obj.widthRectangle)):
+                elif (variable == "widthRectangle" and value in obj.widthRectangle):#np.isin(value, obj.widthRectangle)):
                     self.selected.append(obj)
                 elif (variable == "resultPath" and obj.resultPath == value):
                     self.selected.append(obj)
@@ -93,21 +97,21 @@ class HDF5DB:
                     self.selected.append(obj)
                 elif (variable == "numberOfShapes" and obj.numberOfShapes > value):
                     self.selected.append(obj)
-                elif (variable == "fibreContentCircles" and np.amax(obj.fibreContentCircles) > value):
+                elif (variable == "fibreContentCircles" and np.amin(obj.fibreContentCircles) > value):
                     self.selected.append(obj)
-                elif (variable == "fibreContentRectangles" and np.amax(obj.fibreContentRectangles) > value):
+                elif (variable == "fibreContentRectangles" and np.amin(obj.fibreContentRectangles) > value):
                     self.selected.append(obj)
-                elif (variable == "fibreContentRunners" and np.amax(obj.fibreContentRunners) > value):
+                elif (variable == "fibreContentRunners" and np.amin(obj.fibreContentRunners) > value):
                     self.selected.append(obj)
-                elif (variable == "fvcCircle" and np.amax(obj.fvcCircle) > value):
+                elif (variable == "fvcCircle" and np.amin(obj.fvcCircle) > value):
                     self.selected.append(obj)
-                elif (variable == "radiusCircle" and np.amax(obj.radiusCircle) > value):
+                elif (variable == "radiusCircle" and np.amin(obj.radiusCircle) > value):
                     self.selected.append(obj)
-                elif (variable == "vcRectangle" and np.amax(obj.vcRectangle) > value):
+                elif (variable == "vcRectangle" and np.amin(obj.vcRectangle) > value):
                     self.selected.append(obj)
-                elif (variable == "heightRectangle" and np.amax(obj.heightRectangle) > value):
+                elif (variable == "heightRectangle" and np.amin(obj.heightRectangle) > value):
                     self.selected.append(obj)
-                elif (variable == "widthRectangle" and np.amax(obj.widthRectangle) > value):
+                elif (variable == "widthRectangle" and np.amin(obj.widthRectangle) > value):
                     self.selected.append(obj)
                 elif (variable == "resultPath"):
                     print("The operator " + comparisonOperator + " is not available for resultPath.")
@@ -135,21 +139,21 @@ class HDF5DB:
                     self.selected.append(obj)
                 elif (variable == "numberOfShapes" and obj.numberOfShapes < value):
                     self.selected.append(obj)
-                elif (variable == "fibreContentCircles" and np.amin(obj.fibreContentCircles) < value):
+                elif (variable == "fibreContentCircles" and np.amax(obj.fibreContentCircles) < value):
                     self.selected.append(obj)
-                elif (variable == "fibreContentRectangles" and np.amin(obj.fibreContentRectangles) < value):
+                elif (variable == "fibreContentRectangles" and np.amax(obj.fibreContentRectangles) < value):
                     self.selected.append(obj)
-                elif (variable == "fibreContentRunners" and np.amin(obj.fibreContentRunners) < value):
+                elif (variable == "fibreContentRunners" and np.amax(obj.fibreContentRunners) < value):
                     self.selected.append(obj)
-                elif (variable == "fvcCircle" and np.amin(obj.fvcCircle) < value):
+                elif (variable == "fvcCircle" and np.amax(obj.fvcCircle) < value):
                     self.selected.append(obj)
-                elif (variable == "radiusCircle" and np.amin(obj.radiusCircle) < value):
+                elif (variable == "radiusCircle" and np.amax(obj.radiusCircle) < value):
                     self.selected.append(obj)
-                elif (variable == "vcRectangle" and np.amin(obj.vcRectangle) < value):
+                elif (variable == "vcRectangle" and np.amax(obj.vcRectangle) < value):
                     self.selected.append(obj)
-                elif (variable == "heightRectangle" and np.amin(obj.heightRectangle) < value):
+                elif (variable == "heightRectangle" and np.amax(obj.heightRectangle) < value):
                     self.selected.append(obj)
-                elif (variable == "widthRectangle" and np.amin(obj.widthRectangle) < value):
+                elif (variable == "widthRectangle" and np.amax(obj.widthRectangle) < value):
                     self.selected.append(obj)
                 elif (variable == "resultPath"):
                     print("The operator " + comparisonOperator + " is not available for resultPath.")
@@ -165,7 +169,8 @@ class HDF5DB:
             print("No matches were found for " + str(variable) + " and parameter " + str(value) + ". No filters were applied!")
         else:
             self.HDF5Object = self.selected
-            print("The filter " + str(variable) + " " + str(comparisonOperator) + " " + str(value) + " was applied.")
+            print("The filter " + str(variable) + " " + str(comparisonOperator) + " " + str(value) + " was applied. " + str(len(self.selected)) + " objects were found.")
+            self.hdf5ObjectList =  self.selected
 
     def showSelectionOptions(self):
         self.options = PrettyTable()
@@ -192,17 +197,42 @@ class HDF5DB:
         print(self.options)
 
     def showObjects(self):
-        for val in self.hdf5ObjectList:
-            val.showObjectContent()
+        if not(len(self.hdf5ObjectList) == 0):
+            for val in self.hdf5ObjectList:
+                val.showObjectContent()
+                print("\n")
+        else:
+            print("No objects were found.")
 
-    def save(self):
-        outfile = open("HDF5DB.h5db", "wb" )
-        pickle.dump(self.hdf5ObjectList, outfile)
-        outfile.close()
-        print("HDF5DB saved")
+    def save(self, path, filename="HDF5DB"):
+        if(os.path.exists(path)):
+            if not(len(self.hdf5ObjectList) == 0):
+                if(os.path.isfile(path + "/" + filename + ".h5db")):
+                    print(filename + " will be overwritten. Do you want to continue?\nPlease type in yes to continue.")
+                    userinput = input("")
+                    if not(userinput == "yes"):
+                        print("Nothing has been saved.")
+                        return
+                    else:
+                        print(filename + " will be overwritten.")
+                if not(path == ""):
+                    path = path + "/"                
+                outfile = open(path + filename +  ".h5db", "wb" )
+                pickle.dump(self.hdf5ObjectList, outfile)
+                outfile.close()
+                print("HDF5DB saved")
+            else:
+                print("No objects were found. Nothing was saved!")
+        else:
+            print(path + " does not exist! Nothing was saved!")
 
-    def load(self):
-        infile = open("HDF5DB.h5db", "rb" )
-        self.hdf5ObjectList = pickle.load(infile)
-        infile.close()
-        print("HDF5DB loaded")
+    def load(self, path, filename="HDF5DB"):
+        if not(path == ""):
+            path = path + "/"
+        if(os.path.isfile(path + filename + ".h5db")):
+            infile = open(path + "HDF5DB.h5db", "rb" )
+            self.hdf5ObjectList = pickle.load(infile)
+            infile.close()
+            print("HDF5DB loaded")
+        else:
+            print("There is no h5db-file at the given path with the given name!")
